+++ src/main/java/com/project/LMSSU/domain/Subject.java
@@ -10,14 +10,14 @@ import javax.persistence.Id;
 @NoArgsConstructor
 public class Subject {
     @Id
-    private int subject_id;
+    private Long subject_id;
 
     private String subject_name;
     private String professor_name;
     private String homepage_address;
 
     @Builder
-    public Subject(int subject_id, String subject_name, String professor_name, String homepage_address){
+    public Subject(Long subject_id, String subject_name, String professor_name, String homepage_address){
         this.subject_id = subject_id;
         this.subject_name = subject_name;
         this.professor_name = professor_name;
