+++ src/main/java/com/project/LMSSU/domain/WeeksInfo.java
@@ -13,13 +13,13 @@ import java.time.LocalDate;
 @Data
 public class WeeksInfo {
     @Id
-    private int week;
+    private Long week;
 
     private LocalDate start_date;
     private LocalDate end_date;
 
     @Builder
-    public WeeksInfo(int week, LocalDate start_date, LocalDate end_date){
+    public WeeksInfo(Long week, LocalDate start_date, LocalDate end_date){
         this.week = week;
         this.start_date = start_date;
         this.end_date = end_date;
